move the md5 functions to the library
