+++ Makefile.am
@@ -50,7 +50,7 @@ mutt_SOURCES = account.c addrbook.c address.h alias.c alias.h attach.c \
 	curs_lib.c curs_main.c edit.c editmsg.c enter.c enter_state.h \
 	envelope.h filter.c flags.c format_flags.h from.c getdomain.c group.c \
 	handler.c hash.c hdrline.c header.h headers.c help.c history.c hook.c \
-	init.c keymap.c lib.c list.h main.c mbox.c mbyte.c mbyte_table.h md5.c \
+	init.c keymap.c lib.c list.h main.c mbox.c mbyte.c mbyte_table.h \
 	menu.c mh.c muttlib.c mutt_idna.c mutt_sasl_plain.c mutt_socket.c \
 	mutt_tunnel.c mx.c newsrc.c nntp.c options.h pager.c parameter.c parameter.h \
 	parse.c pattern.c pattern.h pop.c pop_auth.c pop_lib.c postpone.c \
@@ -86,7 +86,7 @@ EXTRA_DIST = account.h attach.h bcache.h browser.h buffer.h buffy.h \
 	ChangeLog.md charset.h CODE_OF_CONDUCT.md compress.h copy.h \
 	COPYRIGHT extlib.c filter.h functions.h gen_defs globals.h \
 	group.h hash.h history.h init.h keymap.h lib.h LICENSE.md mailbox.h \
-	mapping.h mbyte.h md5.h mime.h mime.types mutt.h mutt_commands.h \
+	mapping.h mbyte.h mime.h mime.types mutt.h mutt_commands.h \
 	mutt_curses.h mutt_idna.h mutt_lua.h mutt_menu.h mutt_notmuch.h \
 	mutt_options.h mutt_regex.h mutt_sasl.h mutt_sasl_plain.h \
 	mutt_socket.h mutt_ssl.h mutt_tunnel.h mx.h myvar.h nntp.h OPS \
@@ -98,12 +98,11 @@ EXTRA_DIST = account.h attach.h bcache.h browser.h buffer.h buffy.h \
 
 EXTRA_SCRIPTS =
 
-pgpring_SOURCES = extlib.c lib.c md5.c pgppubring.c sha1.c
+pgpring_SOURCES = extlib.c lib.c pgppubring.c sha1.c
 pgpring_LDADD = $(LIBOBJS) $(NCRYPT_LIBS) $(INTLLIBS) $(LIBLIB)
 pgpring_DEPENDENCIES = $(LIBOBJS) $(NCRYPT_DEPS) $(INTLDEPS) $(LIBLIBDEPS)
 
 mutt_md5_SOURCES = md5.c
-mutt_md5_CFLAGS = -DMD5UTIL
 mutt_md5_LDADD = $(LIBLIB)
 mutt_md5_DEPENDENCIES = $(LIBLIBDEPS)
