+++ imap/auth_cram.c
@@ -31,7 +31,6 @@
 #include "globals.h"
 #include "lib.h"
 #include "lib/lib.h"
-#include "md5.h"
 #include "mutt_socket.h"
 #include "options.h"
 #include "protos.h"
