+++ pgppubring.c
@@ -44,7 +44,6 @@
 #include <unistd.h>
 #include "lib.h"
 #include "lib/lib.h"
-#include "md5.h"
 #include "ncrypt/ncrypt.h"
 #include "ncrypt/pgplib.h"
 #include "ncrypt/pgppacket.h"
