+++ pop_auth.c
@@ -28,12 +28,12 @@
 #include "globals.h"
 #include "lib.h"
 #include "lib/lib.h"
-#include "md5.h"
 #include "mutt_socket.h"
 #include "options.h"
 #include "pop.h"
 #include "protos.h"
 #include "rfc822.h"
+#include "mutt.h"
 #ifdef USE_SASL
 #include <sasl/sasl.h>
 #include <sasl/saslutil.h>
